+++ include/AlgHeap.h
@@ -12,8 +12,8 @@
 
 #include "AlgHeapUtils.h"
 
-int AlgHeap( int n, int *vecX, int *vecY, bool **mtx, int* mina, int* minb, heap h, Pair *stack, unsigned int **hash );
+long long int AlgHeap( int n, long long int *vecX, long long int *vecY, bool **mtx, int* mina, int* minb, heap h, Pair *stack, unsigned int **hash );
 
-int AlgHeap_Prob( int n, int *vecX, int *vecY, const int PROB, heap h, Pair *stack, unsigned int **hash );
+long long int AlgHeap_Prob( int n, long long int *vecX, long long int *vecY, const int PROB, heap h, Pair *stack, unsigned int **hash );
 
 #endif
