+++ src/AlgQT.cpp
@@ -3,17 +3,17 @@
 #include "../include/AlgQT.h"
 #include "../include/Data.h"
 
-#define INF numeric_limits<int>::max()
+#define INF numeric_limits<long long int>::max()
 
 using namespace std;
 
-pair<int, int> AlmostTrivialAlg( int* vecX, int* vecY, bool** resultMtx, int n ) {
+pair<int, int> AlmostTrivialAlg( long long int* vecX, long long int* vecY, bool** resultMtx, int n ) {
 
 	if ( f( resultMtx, 0, 0 ) )
 		return make_pair(0,0);
 
 	int ln = 0, col = 0, maxCol = n;
-	// int best = INF;
+	// long long int best = INF;
 	pair<int, int> best_pair;
 
 	while ( ln < n && maxCol > 0 ) {
@@ -33,13 +33,13 @@ pair<int, int> AlmostTrivialAlg( int* vecX, int* vecY, bool** resultMtx, int n )
 	return best_pair;
 }
 
-pair<int, int> AlmostTrivialAlg_Prob( int* vecX, int* vecY, int n, const int PROB ) {
+pair<int, int> AlmostTrivialAlg_Prob( long long int* vecX, long long int* vecY, int n, const int PROB ) {
 
 	if ( f_prob( PROB ) )
 		return make_pair(0,0);
 
 	int ln = 0, col = 0, maxCol = n;
-	// int best = INF;
+	// long long int best = INF;
 	pair<int, int> best_pair;
 
 	while ( ln < n && maxCol > 0 ) {
